add selected position category course

+++ app/src/main/java/com/kelompoksatuandsatu/preducation/model/CourseViewParam.kt
@@ -33,7 +33,7 @@ data class CourseViewParam(
     val targetAudience: List<String?>?,
     val thumbnail: String?,
     val title: String?,
-    val totalDuration: Int?,
+    val totalDuration: Double?,
     val totalModule: Int?,
     val totalRating: Double?,
     val typeClass: String?
